Move FRR logic into ReflectionResolver
