+++ src/main/java/com/docutools/jocument/GenerationOptionsBuilder.java
@@ -5,6 +5,7 @@ import com.docutools.jocument.formatting.PlaceholderDataFormatter;
 import com.docutools.jocument.formatting.PlaceholderDataFormattingOption;
 import com.docutools.jocument.image.DefaultImageStrategy;
 import com.docutools.jocument.image.ImageStrategy;
+import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
@@ -27,6 +28,7 @@ public final class GenerationOptionsBuilder {
   private ImageStrategy imageStrategy;
   private final List<PlaceholderDataFormattingOption> formattingOptions = new ArrayList<>();
   private BiFunction<String, Locale, Optional<String>> translationFunction = null;
+  private Duration waitTime = Duration.ofSeconds(30);
 
   public GenerationOptionsBuilder() {
     this.imageStrategy = DefaultImageStrategy.instance();
@@ -42,7 +44,7 @@ public final class GenerationOptionsBuilder {
   }
 
   public <T> GenerationOptionsBuilder format(Class<T> filter, LocalisedPlaceholderDataFormatter<T> formatter) {
-    formattingOptions.add(new PlaceholderDataFormattingOption<T>(obj -> obj.getClass().isAssignableFrom(filter), formatter));
+    formattingOptions.add(new PlaceholderDataFormattingOption<>(obj -> obj.getClass().isAssignableFrom(filter), formatter));
     return this;
   }
 
@@ -51,8 +53,13 @@ public final class GenerationOptionsBuilder {
     return this;
   }
 
+  public GenerationOptionsBuilder withMaximumWaitTime(Duration waitTime) {
+    this.waitTime = waitTime;
+    return this;
+  }
+
   public GenerationOptions build() {
-    return new GenerationOptions(imageStrategy, formattingOptions, translationFunction);
+    return new GenerationOptions(imageStrategy, waitTime, formattingOptions, translationFunction);
   }
 
 }
